requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3

black==22.3.0
flake8==4.0.1

pytest-django==4.5.2 
pytest==7.1.2